+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: tlavelle <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/05/08 15:12:15 by tlavelle          #+#    #+#             */
-/*   Updated: 2020/05/26 14:21:44 by tlavelle         ###   ########.fr       */
+/*   Updated: 2020/05/28 13:29:14 by tlavelle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int count;
+	size_t count;
 
 	count = 0;
 	while (s[count] != '\0')
